combine orgsite and site django apps together
these two apps were so dependant on each other that there was not discernable boundary
