+++ server/haztrak/settings/base.py
@@ -56,7 +56,6 @@ INSTALLED_APPS = [
     "manifest",
     "wasteline",
     "org",
-    "orgsite",
     "profile",
 ]
 
@@ -236,6 +235,6 @@ GUARDIAN_MONKEY_PATCH = False
 TRAK_ORG_MODEL = "org.Org"
 TRAK_RCRAINFO_SITE_MODEL = "rcrasite.RcraSite"
 TRAK_MANIFEST_MODEL = "manifest.Manifest"
-TRAK_SITE_MODEL = "orgsite.Site"
+TRAK_SITE_MODEL = "org.Site"
 TRAK_WASTELINE_MODEL = "wasteline.Wasteline"
 TRAK_HANDLER_MODEL = "handler.Handler"
