+++ server/org/admin.py
@@ -3,8 +3,15 @@ from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
 from guardian.admin import GuardedModelAdmin
 
-from org.models import Org, OrgAccess, OrgGroupObjectPermission, OrgUserObjectPermission
-from orgsite.models import Site
+from org.models import (
+    Org,
+    OrgAccess,
+    OrgGroupObjectPermission,
+    OrgUserObjectPermission,
+    Site,
+    SiteGroupObjectPermission,
+    SiteUserObjectPermission,
+)
 
 admin.site.register(OrgAccess)
 
@@ -35,6 +42,30 @@ class OrgUserObjectPermissionAdmin(GuardedModelAdmin):
 
 @admin.register(OrgGroupObjectPermission)
 class OrgGroupObjectPermissionAdmin(GuardedModelAdmin):
+    def formfield_for_foreignkey(self, db_field, request, **kwargs):
+        if db_field.name == "permission":
+            content_type = ContentType.objects.get_for_model(Site)
+            kwargs["queryset"] = Permission.objects.filter(content_type=content_type)
+        return super().formfield_for_foreignkey(db_field, request, **kwargs)
+
+
+@admin.register(Site)
+class HaztrakSiteAdmin(admin.ModelAdmin):
+    list_display = ["__str__", "last_rcrainfo_manifest_sync"]
+    search_fields = ["rcra_site__epa_id"]
+
+
+@admin.register(SiteUserObjectPermission)
+class SiteUserObjectPermissionAdmin(GuardedModelAdmin):
+    def formfield_for_foreignkey(self, db_field, request, **kwargs):
+        if db_field.name == "permission":
+            content_type = ContentType.objects.get_for_model(Site)
+            kwargs["queryset"] = Permission.objects.filter(content_type=content_type)
+        return super().formfield_for_foreignkey(db_field, request, **kwargs)
+
+
+@admin.register(SiteGroupObjectPermission)
+class SiteGroupObjectPermissionAdmin(GuardedModelAdmin):
     def formfield_for_foreignkey(self, db_field, request, **kwargs):
         if db_field.name == "permission":
             content_type = ContentType.objects.get_for_model(Site)
