+++ server/org/serializers.py
@@ -1,7 +1,8 @@
 from rest_framework import serializers
 from rest_framework.serializers import ModelSerializer
 
-from org.models import Org
+from org.models import Org, Site
+from rcrasite.serializers import RcraSiteSerializer
 
 
 class OrgSerializer(ModelSerializer):
@@ -25,3 +26,20 @@ class OrgSerializer(ModelSerializer):
             "id",
             "rcrainfoIntegrated",
         ]
+
+
+class SiteSerializer(ModelSerializer):
+    """
+    Haztrak Site model serializer for JSON marshalling/unmarshalling
+    """
+
+    name = serializers.CharField(
+        required=False,
+    )
+    handler = RcraSiteSerializer(
+        source="rcra_site",
+    )
+
+    class Meta:
+        model = Site
+        fields = ["name", "handler"]
