+++ server/org/tests/conftest.py
@@ -2,8 +2,7 @@ from typing import Optional
 
 import pytest
 
-from org.models import Org
-from orgsite.models import Site
+from org.models import Org, Site
 from rcrasite.models import RcraSite
